[FEAT] Enhance memory web app configuration
Detailed Changes:
- Added support for configuring the port using the `PORT` environment variable
- Changed the default port from 5000 to 5500

Impact: This change allows the memory web app to be easily configured to run on a different port, which can be useful when deploying the application.

Additional Notes: The `PORT` environment variable is a common convention for specifying the port for web applications, especially when deploying to cloud platforms.

Related Issues: [<slug>](https://go/issues/<slug>)